+++ picopt/config/__init__.py
@@ -50,7 +50,7 @@ _TEMPLATE = MappingTemplate(
                 "symlinks": bool,
                 "timestamps": bool,
                 "timestamps_check_config": bool,
-                "timestamps_use_archive": bool,
+                "timestamps_ignore_archive_entry_mtimes": bool,
                 "verbose": Integer(),
                 "computed": Optional(
                     MappingTemplate(