+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## v5.3.2
 
+- Add option --timestamps-ignore-archive-entry-mtimes.
 - Add type hints and py.typed sentinel file.
 
 ## v5.3.1